refactor(exposition): replace `fetch.redirect` option with standalone directive
