+++ extensions/exposition/schemas/octets/fetch.cos.yaml
@@ -1,4 +1,3 @@
 blob: boolean
 meta: boolean
-redirect: string
 _: true
