+++ extensions/exposition/source/Directive.ts
@@ -92,5 +92,7 @@ export const shortcuts: RTD.syntax.Shortcuts = new Map([
   ['rule', 'auth:rule'],
   ['incept', 'auth:incept'],
   ['input', 'io:input'],
-  ['output', 'io:output']
+  ['output', 'io:output'],
+  ['fetch', 'octets:fetch'],
+  ['redirect', 'octets:redirect']
 ])
