+++ extensions/exposition/source/directives/octets/Fetch.ts
@@ -1,45 +1,35 @@
 import { posix } from 'node:path'
-import { Readable } from 'node:stream'
 import { Forbidden, NotFound } from '../../HTTP'
 import * as schemas from './schemas'
 import { Directive } from './Directive'
-import type { Remotes } from '../../Remotes'
+import type { Readable } from 'node:stream'
 import type { Maybe } from '@toa.io/types'
 import type { Entry } from '@toa.io/extensions.storages'
 import type { Component } from '@toa.io/core'
 import type { Output } from '../../io'
 import type { Input } from './types'
-import type { Parameter } from '../../RTD'
 
 export class Fetch extends Directive {
   public readonly targeted = true
 
   private readonly options: Required<Options> = {
     blob: true,
-    meta: false,
-    redirect: null
+    meta: false
   }
 
   private readonly discovery: Promise<Component>
   private storage!: Component
-  private readonly remotes: Remotes
-  private connecting: Promise<Component> | null = null
-  private remote!: Component
 
-  public constructor (options: Options | null, discovery: Promise<Component>, remotes: Remotes) {
+  public constructor (options: Options | null, discovery: Promise<Component>) {
     super()
 
     schemas.fetch.validate(options)
     Object.assign(this.options, options)
 
     this.discovery = discovery
-    this.remotes = remotes
   }
 
-  public async apply (storage: string, input: Input, parameters: Parameter[]): Promise<Output> {
-    if (this.options.redirect !== null)
-      return this.redirect(input, parameters)
-
+  public async apply (storage: string, input: Input): Promise<Output> {
     this.storage ??= await this.discovery
 
     const variant = posix.basename(input.request.url).includes('.')
@@ -83,48 +73,6 @@ export class Fetch extends Directive {
     }
   }
 
-  private async redirect (input: Input, parameters: Parameter[]): Promise<Output> {
-    if ('if-none-match' in input.request.headers)
-      return { status: 304 }
-
-    const [operation, name, namespace = 'default'] = this.options.redirect!.split('.').reverse()
-
-    if (this.connecting === null)
-      this.connecting = this.remotes.discover(namespace, name)
-
-    this.remote ??= await this.connecting
-
-    const url = await this.remote.invoke<Maybe<string>>(operation, {
-      input: {
-        authority: input.authority,
-        path: input.request.url,
-        parameters: Object.fromEntries(parameters.map(({ name, value }) => [name, value]))
-      }
-    })
-
-    if (url instanceof Error)
-      throw new NotFound(url)
-
-    const response = await fetch(url)
-
-    if (!response.ok)
-      throw new NotFound()
-
-    const headers = new Headers()
-
-    for (const header of ['content-type', 'content-length', 'etag']) {
-      const value = response.headers.get(header)
-
-      if (value !== null)
-        headers.set(header, value)
-    }
-
-    return {
-      headers,
-      body: response.body === null ? null : Readable.fromWeb(response.body as any)
-    }
-  }
-
   private async get (storage: string, input: Input): Promise<Output> {
     const entry = await this.storage.invoke<Maybe<Entry>>('get', {
       input: {
@@ -143,7 +91,6 @@ export class Fetch extends Directive {
 export interface Options {
   blob?: boolean
   meta?: boolean
-  redirect?: string | null
 }
 
 interface FetchResult {
