+++ extensions/exposition/source/directives/octets/Octets.ts
@@ -4,6 +4,7 @@ import { Store } from './Store'
 import { Fetch } from './Fetch'
 import { List } from './List'
 import { Delete } from './Delete'
+import { Redirect } from './Redirect'
 import { WorkflowDirective } from './Workflow'
 import type { Directive } from './Directive'
 import type { Output } from '../../io'
@@ -65,6 +66,7 @@ const DIRECTIVES: Record<string, Constructor> = {
   fetch: Fetch,
   list: List,
   delete: Delete,
+  redirect: Redirect,
   workflow: WorkflowDirective
 }
